% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KoreanBaseballNews-package.R
\docType{package}
\name{KoreanBaseballNews-package}
\alias{KoreanBaseballNews}
\alias{KoreanBaseballNews-package}
\title{KoreanBaseballNews: To provide the Korean Baseball news inforamtion}
\description{
This data package contains Korean Baseball news information.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/MrKevinNa/KoreanBaseballNews}
  \item Report bugs at \url{https://github.com/MrKevinNa/KoreanBaseballNews/issues}
}

}
\author{
\strong{Maintainer}: DrKevin Na \email{drkevin2022@gmail.com}

}
\keyword{internal}
