% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/koreanbaseballnews2018.R
\docType{data}
\name{KoreanBaseballNews2018}
\alias{KoreanBaseballNews2018}
\title{Korean Baseball News in year of 2018}
\format{tibble with columns}
\source{
https://sports.news.naver.com/kbaseball/news/index.nhn?isphoto=N
}
\usage{
KoreanBaseballNews2018
}
\description{
Korean Baseball News in year of 2018
}
\examples{
  KoreanBaseballNews2018
}
\keyword{datasets}
